Final split (for now) into helpers
