+++ src/helpers/miscHelpers.ts
@@ -1,0 +1,6 @@
+type AnyFunction = (...params: any[]) => any;
+
+const composeFn = (...fns: AnyFunction[]) => (...args: any[]) =>
+  fns.forEach(fn => fn && fn(...args));
+
+export { composeFn };
